Change of finger datatype to int + testing of computation of gyro

+++ hardware/SignSaya/SignSaya/fingers.h
@@ -36,7 +36,7 @@ public:
     pinMode(pinNumber, INPUT);
   }
 
-  float read() {
+  int read() {
     currentValue = mapRead();
     return movingAverage(currentValue);
   }
